Made React Components Feel more Responsive
Components like DonorForm and DonationInfo felt sluggish when you typed, and that was because the entire checkout page was getting re rendered on every key press.

i solved this by moving the state of the donor and donation info from the parent component to the DonorForm and DonationInfo components respectively.

that way, only the component itself re renders when its own state changes, and not the entire page.

This local state made us lose the ability to update the parent Checkout component's state from the child component, but we can fix that by passing a useRef from the parent component to the child component. Now that the parent component has a reference to the child component, it can call the child component's methods to update its state. We only ever update the Parent's donor and donation state when the user clicks the next button, so we don't need to worry about the parent's state getting out of sync with the child's state.

Now we get the best of both worlds: responsiveness + global state from the parent component.