TODO: Share the same DonationInfo Component
For the time being, the Donation Input and Info is not visible in the mobile sizes (whatever that measure is im still not sure from looking at the code). That's because we moved the Mobile Donation Info into the Grid with id="desktop-donation-info", and that whole grid dissappears when the screen size becomes xs in its display {xs: 'none'} property.

Now we have tgo figure out how too
Just Change the CSS attributes +visibility Depending on Desktop or Mobile.