Environment Variables for a build:test Script in package.json
build:test is basically build but its environment variables are hosted in .env.test

Also, a required-env.js file checks to see if we have the required environment variables

We just have to make sure that different environments have different Gateway logins from the QGWdb Engine.

Production logs in to real account,
Development and Test Log in to test accounts.