QuantumGateway Transparent DB Engine: https://www.quantumgateway.com/view_developer.php?Cat1=3
for the time being, we're just using local host as our POST URL so that we can have server to server communication.

In local environment you have to run a local server to process transactions that use transparent quantum gateway db engine. we're gonna have to build and deploy our own server. Here's the code:
https://github.com/pjfitacc/qdbwserver