Recurring Input Form Fields
Included so far:
1. initialIntervalAmount (number)
2. recurAmount (number)
3. overrideRecurPrice

overrideRecurPrice is set to "Y". Indicating that the recurAmount will not be the default recipe ID, but it will be recurAmount from now on.

initialIntervalAmount is null in its initial state so its first value when rendered in the RecurringInput form is the regular amount field in formStore. Otherwise, it'll take the initialIntervalAmount from then on.

Other required recurring input form fields:
1. recipeID - select component - use the RecurringRecipes in models.js
2. timesToRecur - number
3. overrideRecurDay - "Y"/"N" if the donor wants to override the default RecurringRecipes cycleDay to today.